+++ tests/libYARP_sig/MatrixTest.cpp
@@ -416,8 +416,8 @@ public:
         std::string s = writer.toString();
         Bottle bot;
         bot.fromBinary(s.c_str(),(int)s.length());
-        checkEqual(bot.get(0).asInt32(),rr,"row count matches");
-        checkEqual(bot.get(1).asInt32(),cc,"column count matches");
+        checkEqual((unsigned int) bot.get(0).asInt32(),rr,"row count matches");
+        checkEqual((unsigned int) bot.get(1).asInt32(),cc,"column count matches");
         Bottle *lst = bot.get(2).asList();
         checkTrue(lst!=nullptr,"have data");
         if (!lst) return;
@@ -464,8 +464,8 @@ public:
         Bottle *bot2 = bot.get(1).asList();
         checkTrue(bot1!=nullptr&&bot2!=nullptr,"got head/body");
         if (bot1==nullptr || bot2==nullptr) return;
-        checkEqual(bot1->get(0).asInt32(),rr,"row count matches");
-        checkEqual(bot1->get(1).asInt32(),cc,"column count matches");
+        checkEqual((unsigned int) bot1->get(0).asInt32(),rr,"row count matches");
+        checkEqual((unsigned int) bot1->get(1).asInt32(),cc,"column count matches");
         Bottle *lst = bot1->get(2).asList();
         checkTrue(lst!=nullptr,"have data");
         if (!lst) return;
